backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.125.0",
    "landingai-ade>=1.2.0",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.21",
    "sqlmodel>=0.0.27",
    "supabase>=2.27.0",
    "uvicorn>=0.38.0",
]
